Changes based on review of cycleThroughPhases - change times to milliseconds, return type to double

+++ src/TrafficLight.h
@@ -55,7 +55,7 @@ public:
 private:
     // typical behaviour methods
     void cycleThroughPhases();
-    int calculateCycleLength();
+    double calculateCycleLength();
 
     // FP.4b : create a private member of type MessageQueue for messages of type TrafficLightPhase 
     // and use it within the infinite loop to push each new TrafficLightPhase into it by calling
